pyproject.toml
[project]
name = "freshdesk-mbox-importer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pydantic-settings>=2.10.1",
]
